added console read key

+++ tumakov2/structs/bankStruct.cs
@@ -1,11 +1,10 @@
 ﻿using bankEnum;
-using System.Data;
 namespace bankStruct
 {
     public struct bank
     {
         public string accType ;
-        public decimal balance;
+        public double balance;
         public int id;
     }
 }
